Fix back button that wouldn't work once entering the Sign Up page
This fixes the case where users aren't logged in and search for a domain. Upon clicking a specific suggestion they are redirected to the signup when the Checkout component is mounted. This makes it impossible to get back to the previous page (i.e. the Search page) because of a loop.