Fix submission of Login and Signup pages when navigating back-and-forth
This fix a regression that prevents users from submitting the form on the Login or the Signup page more than once. To be more specific, this happens when users access any of these pages for the first time and submit the form. They are then redirected to the Verify User page. If they navigate back to the previous page they won't be able to submit the form again - clicking on the submit button doesn't do anything.