Contact Information: add form validation (#151)
* Validate contact information via the API

* Update Redux Form package to latest version

* Rename identifier of postal code field in Contact Information page

This makes sure the field's identifier matches the one of the data returned by the API.

* Use `redux-form` actions in `validateContactInformation`

This lets us associate these with the form in `state.form.contact-information`

* Normalize errors when validating contact information

* Normalize payload for validating contact information

* Add tests for normalization functions

* Rename functions that camelize or snakeify object keys to be more explicit

* Retrieves list of countries supported for domain registrations

We were pulling the wrong list of countries.

* Split the name field on the Contact Information page

This adds two fields instead: one for the first name and another one for the last name. That's what the WordPress.com API as well as other third party APIs expect, and it will make our code way simpler.