Fix text flashing because of wrong font
This fixes a problem with quotes that are escaped when styles are rendered server-side. This broke the font family name which resulted in displaying text with one font when the page loads and then with another one when the javascript bundle is loaded and injects styles.

Since apparently quotes are not mandatory for font family names that contain spaces this simply fixes the aforementioned problem by removing them.